+++ CHANGELOG.md
@@ -13,7 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Non-deterministic tests are run multiple times similar to {Quick,Scala}check (#786)
 - `quint run` returns a non-zero exit code on invariant violation and errors (#793)
 - Support for imports from files: `import ... from <filename>` (#800)
+- suppress comments in REPL (#806)
 - `quint repl` is printing the same version number as returned by `quint --version` (#804)
+- add the command `.seed` in REPL (#812)
+- fix `quint run` to output compile errors again (#812)
 
 ### Changed
 
@@ -22,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+
+- Proper errors are now reported when a lambda returns an operator (#811)
+
 ### Security
 
 ## v0.9.1 -- 2023-04-04